Book DataLayer Complete
Took some time to figure out how to use the json data as source, since
VS2015 refuse to accept it as a database. (XML is ok, but not JSON. wtf)

After a search for third party solution I found out that Newtonsoft Json
package already was installed.

The datalayer have an improved search method that sorts the result after
a relevance score.